modified the backend token capturing logic now instead of capuring in apollo-client I capture token and add to headers  in middleware.
