--- src/app/middleware.ts
@@ -1,22 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-
-export async function middleware(req: NextRequest) {
-  const token = req.cookies.get("token")?.value;
-
-  if (!token) {
-    return NextResponse.redirect(new URL("/", req.url));
-  }
-
-  const requestHeaders = new Headers(req.headers);
-  requestHeaders.set("Authorization", `Bearer ${token}`);
-
-  return NextResponse.next({
-    request: {
-      headers: requestHeaders,
-    },
-  });
-}
-
-export const config = {
-  matcher: "/api",
-};
